Updated all network code to 1.1.25, still not stable

+++ src/me/woder/network/SoundEffect41.java
@@ -16,7 +16,8 @@ public class SoundEffect41 extends Packet{
         buf.readInt();
         buf.readInt();
         buf.readFloat();
-        buf.readByte();
+        buf.readFloat();
+        //buf.readByte();
     }
 
 }
